+++ src/crypto/wincrypto/cert.rs
@@ -1,17 +1,17 @@
 use crate::crypto::dtls::DTLS_CERT_IDENTITY;
 use crate::crypto::Fingerprint;
 use std::sync::Arc;
-use str0m_wincrypto::{Certificate, WinCryptoError};
+use str0m_wincrypto::{WinCryptoCertificate, WinCryptoError};
 
 #[derive(Clone, Debug)]
 pub struct WinCryptoDtlsCert {
-    pub(crate) certificate: Arc<Certificate>,
+    pub(crate) certificate: Arc<WinCryptoCertificate>,
 }
 
 impl WinCryptoDtlsCert {
     pub fn new() -> Self {
         let certificate = Arc::new(
-            Certificate::new_self_signed(&format!("CN={}", DTLS_CERT_IDENTITY))
+            WinCryptoCertificate::new_self_signed(&format!("CN={}", DTLS_CERT_IDENTITY))
                 .expect("Failed to create self-signed certificate"),
         );
         Self { certificate }
@@ -23,7 +23,7 @@ impl WinCryptoDtlsCert {
 }
 
 pub(super) fn create_fingerprint(
-    certificate: &str0m_wincrypto::Certificate,
+    certificate: &str0m_wincrypto::WinCryptoCertificate,
 ) -> Result<Fingerprint, WinCryptoError> {
     certificate.sha256_fingerprint().map(|bytes| Fingerprint {
         hash_func: "sha-256".into(),
