+++ src/crypto/wincrypto/sha1.rs
@@ -1,4 +1,4 @@
-use str0m_wincrypto::sha1_hmac as wincrypto_sha1_hmac;
+use str0m_wincrypto::wincrypto_sha1_hmac;
 
 pub fn sha1_hmac(key: &[u8], payloads: &[&[u8]]) -> [u8; 20] {
     match wincrypto_sha1_hmac(key, payloads) {
