+++ wincrypto/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 thiserror = { version = "1.0.38" }
+tracing = "0.1.37"
 windows = { version = "0.58", features=["Win32_Security_Cryptography", "Win32_Security_Authentication_Identity", "Win32_Security_Credentials",]}
 windows-core = { version = "0"}
 windows-strings = { version = "0.1"}
