+++ wincrypto/src/cert.rs
@@ -4,11 +4,11 @@ use windows_strings::PSTR;
 use super::WinCryptoError;
 
 #[derive(Debug)]
-pub struct Certificate(pub(crate) *const CERT_CONTEXT);
-unsafe impl Send for Certificate {}
-unsafe impl Sync for Certificate {}
+pub struct WinCryptoCertificate(pub(crate) *const CERT_CONTEXT);
+unsafe impl Send for WinCryptoCertificate {}
+unsafe impl Sync for WinCryptoCertificate {}
 
-impl Certificate {
+impl WinCryptoCertificate {
     pub fn new_self_signed(subject: &str) -> Result<Self, WinCryptoError> {
         unsafe {
             let subject = HSTRING::from(subject);
@@ -114,13 +114,13 @@ impl Certificate {
     }
 }
 
-impl From<*const CERT_CONTEXT> for Certificate {
+impl From<*const CERT_CONTEXT> for WinCryptoCertificate {
     fn from(value: *const CERT_CONTEXT) -> Self {
         Self(value)
     }
 }
 
-impl Drop for Certificate {
+impl Drop for WinCryptoCertificate {
     fn drop(&mut self) {
         unsafe {
             _ = CertFreeCertificateContext(Some(self.0));
@@ -133,7 +133,7 @@ mod tests {
     #[test]
     fn verify_self_signed() {
         unsafe {
-            let cert = super::Certificate::new_self_signed("cn=WebRTC").unwrap();
+            let cert = super::WinCryptoCertificate::new_self_signed("cn=WebRTC").unwrap();
 
             // Verify it is self-signed.
             let subject = (*(*cert.0).pCertInfo).Subject;
@@ -146,7 +146,7 @@ mod tests {
 
     #[test]
     fn verify_fingerprint() {
-        let cert = super::Certificate::new_self_signed("cn=WebRTC").unwrap();
+        let cert = super::WinCryptoCertificate::new_self_signed("cn=WebRTC").unwrap();
         let fingerprint = cert.sha256_fingerprint().unwrap();
         assert_eq!(fingerprint.len(), 32);
     }
